:recycle: Refactored Favorite Movies Protocol
