+++ ios-challenge-moviedb/ios-challenge-moviedb/Presenter/Movie/MoviePresenter.swift
@@ -65,9 +65,7 @@ class MoviePresenter {
         let selectedMovie = movies[index]
         delegate?.selectedMovie(movie: selectedMovie)
     }
-}
-
-extension MoviePresenter: FavoriteMoviesProtocol {
+    
     func isMovieFavorite(movie: Movie) -> Bool {
         let favoriteMovies = LocalData.object.getAllFavoriteMovies()
         if favoriteMovies[movie.id] == nil {
@@ -75,21 +73,17 @@ extension MoviePresenter: FavoriteMoviesProtocol {
         }
         return true
     }
-    
-    func handleMovieFavorite(movie: Movie) {
+
+    func handleMovieFavoriteTap(movie: Movie) {
         if isMovieFavorite(movie: movie) == false {
             LocalData.object.makeMovieFavorite(movie: movie)
         } else {
             LocalData.object.makeMovieNotFavorite(movie: movie)
         }
     }
-    
-    func changeButtonImage(button: UIButton, movie: Movie) {
+
+    func handleChangeButtonImage(movie: Movie) -> Bool {
         let isFavorite = isMovieFavorite(movie: movie)
-        if isFavorite == true {
-            button.setImage(UIImage(named: Constants.FavoriteButton.imageNamedFull), for: .normal)
-        } else {
-            button.setImage(UIImage(named: Constants.FavoriteButton.imageNamedNormal), for: .normal)
-        }
+        return isFavorite
     }
 }
