+++ ios-challenge-moviedb/ios-challenge-moviedb/View/Data Sources/FavoritesCollectionViewDataSource.swift
@@ -31,7 +31,7 @@ class FavoritesCollectionViewDataSource: NSObject, UICollectionViewDataSource {
             cell.movieImage.kf.indicatorType = .activity
             let movie = movies[indexPath.item]
             cell.movieTitle.text = movie.title
-            viewController?.presenter?.changeButtonImage(button: cell.favoriteButton, movie: movie)
+            viewController?.changeButtonImage(button: cell.favoriteButton, movie: movie)
             cell.favoriteButton.tag = movie.id
             cell.favoriteButton.addTarget(self, action: #selector(handleFavorite(_:)), for: .touchUpInside)
             let moviePosterImageURL = viewController?.presenter?.getMovieImageURL(width: 200, path: movie.posterPath ?? "")
@@ -57,7 +57,7 @@ class FavoritesCollectionViewDataSource: NSObject, UICollectionViewDataSource {
             }
         }
         guard let movie = cellMovie else { return }
-        viewController?.presenter?.handleMovieFavorite(movie: movie)
-        viewController?.presenter?.changeButtonImage(button: sender, movie: movie)
+        viewController?.handleMovieFavorite(movie: movie)
+        viewController?.changeButtonImage(button: sender, movie: movie)
      }
 }
