+++ ios-challenge-moviedb/ios-challenge-moviedb/View/Favorite/FavoriteViewController.swift
@@ -158,7 +158,6 @@ extension FavoriteViewController: UISearchBarDelegate {
         } else {
             checkIfNeedsToRemoveError()
         }
-        
         reloadData()
     }
     
@@ -169,5 +168,17 @@ extension FavoriteViewController: UISearchBarDelegate {
     }
 }
 
-
-
+extension FavoriteViewController : FavoriteMoviesProtocol {
+    func handleMovieFavorite(movie: Movie) {
+        presenter?.handleMovieFavoriteTap(movie: movie)
+    }
+    
+    func changeButtonImage(button: UIButton, movie: Movie) {
+        let isFavorite = presenter?.handleChangeButtonImage(movie: movie)
+        if isFavorite == true {
+            button.setImage(UIImage(named: Constants.FavoriteButton.imageNamedFull), for: .normal)
+        } else {
+            button.setImage(UIImage(named: Constants.FavoriteButton.imageNamedNormal), for: .normal)
+        }
+    }
+}
